1. Fold CompilerInstanceWithContext into DependencyScanningWorker.cpp
