+++ clang/include/clang/Tooling/DependencyScanningTool.h
@@ -104,9 +104,6 @@ public:
   /// Given a compilation context specified via the Clang driver command-line,
   /// gather modular dependencies of module with the given name, and return the
   /// information needed for explicit build.
-  /// TODO: this method should be removed as soon as Swift and our C-APIs adopt
-  /// CompilerInstanceWithContext. We are keeping it here so that it is easier
-  /// to coordinate with Swift and C-API changes.
   llvm::Expected<dependencies::TranslationUnitDeps> getModuleDependencies(
       StringRef ModuleName, ArrayRef<std::string> CommandLine, StringRef CWD,
       const llvm::DenseSet<dependencies::ModuleID> &AlreadySeen,
